STOR-6 refactor packages
