+++ pom.xml
@@ -66,17 +66,13 @@
             <artifactId>spring-data-jpa</artifactId>
             <version>2.6.0</version>
         </dependency>
+
         <dependency>
             <groupId>jakarta.persistence</groupId>
             <artifactId>jakarta.persistence-api</artifactId>
             <version>3.0.0</version>
         </dependency>
 
-        <dependency>
-            <groupId>org.hibernate.validator</groupId>
-            <artifactId>hibernate-validator</artifactId>
-            <version>6.0.16.Final</version>
-        </dependency>
     </dependencies>
 
     <build>
