+++ src/main/java/com/butbetter/storage/file_upload/Exceptions/StorageFileNotFoundException.java
@@ -1,4 +1,4 @@
-package com.butbetter.storage.FileUpload.Exceptions;
+package com.butbetter.storage.file_upload.Exceptions;
 
 public class StorageFileNotFoundException extends Throwable {
