+++ src/main/java/com/butbetter/storage/file_upload/FileStorageService.java
@@ -1,10 +1,10 @@
-package com.butbetter.storage.FileUpload;
+package com.butbetter.storage.file_upload;
 
-import com.butbetter.storage.CSV.CSVImportService;
-import com.butbetter.storage.CSV.Exceptions.FaultyCSVException;
-import com.butbetter.storage.FileUpload.Exceptions.StorageException;
-import com.butbetter.storage.FileUpload.Exceptions.StorageFileNotFoundException;
-import com.butbetter.storage.FileUpload.Properties.StorageProperties;
+import com.butbetter.storage.csv.CSVImportService;
+import com.butbetter.storage.csv.Exceptions.FaultyCSVException;
+import com.butbetter.storage.file_upload.Exceptions.StorageException;
+import com.butbetter.storage.file_upload.Exceptions.StorageFileNotFoundException;
+import com.butbetter.storage.file_upload.Properties.StorageProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
