+++ src/main/java/com/butbetter/storage/file_upload/StorageService.java
@@ -1,8 +1,8 @@
-package com.butbetter.storage.FileUpload;
+package com.butbetter.storage.file_upload;
 
-import com.butbetter.storage.CSV.Exceptions.FaultyCSVException;
-import com.butbetter.storage.FileUpload.Exceptions.StorageException;
-import com.butbetter.storage.FileUpload.Exceptions.StorageFileNotFoundException;
+import com.butbetter.storage.csv.Exceptions.FaultyCSVException;
+import com.butbetter.storage.file_upload.Exceptions.StorageException;
+import com.butbetter.storage.file_upload.Exceptions.StorageFileNotFoundException;
 import org.springframework.core.io.Resource;
 import org.springframework.web.multipart.MultipartFile;
