+++ src/main/java/com/butbetter/storage/model/ProductInformation.java
@@ -5,12 +5,11 @@ import org.hibernate.validator.constraints.Range;
 
 import javax.persistence.*;
 import java.io.Serializable;
-import com.butbetter.storage.customConverter.BeanAddressConverter;
-import com.butbetter.storage.customConverter.BeanOffsetDateTimeConverter;
+import com.butbetter.storage.custom_converter.BeanAddressConverter;
+import com.butbetter.storage.custom_converter.BeanOffsetDateTimeConverter;
 import com.opencsv.bean.CsvBindByName;
 import com.opencsv.bean.CsvCustomBindByName;
 
-import javax.persistence.*;
 import java.time.OffsetDateTime;
 import java.util.Objects;
 import java.util.UUID;
