+++ src/test/java/com/butbetter/storage/csv/CSVConverterTest.java
@@ -1,6 +1,6 @@
-package com.butbetter.storage.CSV;
+package com.butbetter.storage.csv;
 
-import com.butbetter.storage.FileUpload.Exceptions.StorageFileNotFoundException;
+import com.butbetter.storage.file_upload.Exceptions.StorageFileNotFoundException;
 import com.butbetter.storage.model.Address;
 import com.butbetter.storage.model.ProductInformation;
